Fix the drawable resource can't be loaded from RRO package
Since refer to resources from the overlay package using the IDs defined
in the original package, NotFoundException occurs and icons can NOT be
loaded.
These are caused by using Icon.createWithResource(Resources, ResId),
and usage of this API is NOT supported.

To resolve this issue,  use Icon.createWithResource(Context, ResId) for
create icons with original package.

Test: manual - Check that SIP call account icon is displayed with the
specified in the overlay package
Bug: 124352738

Change-Id: I43be48818b190698a4344d8fc8f2ac5a6927702a
Signed-off-by: DennySPB <dennyspb@gmail.com>